Updated fmri_timeseries and segment.h to save averages to a 2D image.
Changed the segment function to return the number of labels to be
processed. Changed fmri_timeseries to save the resuts of the averaging
of each label to a 2D image (where the label is on the faster axis
and time is the slower. Note, I will need to add some metadata for
the labels.