Fixes for reweighting, extra RegularizedResampler function, more output, BoldModel Bugfix, removed warnings.
When a particle gets weighted to 0, it now gets a "default" value that
won't go off the deep end.

Added a function that will accept a covariance to bet given when
resampling rather than having it be calculated by the resampling
algorithm itself. This could permit pre-stratified resampled values for
the covariance to be used in cases of extreme particle impoverishment.

Boldmodel had a problem where arguments were out of order. Misc warnings